Move document-info components
