--- src/components/editor/document-bar/document-info-time-line-helper/unitalic-bold-text.scss
@@ -1,3 +1,0 @@
-.font-style-normal {
-  font-style: normal;
-}
