+++ src/components/editor/document-bar/document-info/document-info-time-line.scss
