+++ src/components/editor/document-bar/document-info/document-info-time-line.tsx
@@ -1,11 +1,11 @@
 import { Moment } from 'moment'
 import React from 'react'
 import { Trans, useTranslation } from 'react-i18next'
-import { IconName } from '../../common/fork-awesome/fork-awesome-icon'
+import { IconName } from '../../../common/fork-awesome/fork-awesome-icon'
 import { DocumentInfoLine } from './document-info-line'
 import './document-info-time-line.scss'
-import { TimeFromNow } from './document-info-time-line-helper/time-from-now'
-import { UserAvatar } from '../../common/user-avatar/user-avatar'
+import { TimeFromNow } from './time-from-now'
+import { UserAvatar } from '../../../common/user-avatar/user-avatar'
 
 export interface DocumentInfoLineWithTimeProps {
   time: Moment,
