+++ src/components/editor/document-bar/document-info/time-from-now.tsx
@@ -1,11 +1,11 @@
 import { Moment } from 'moment'
 import React from 'react'
 
-export interface ItalicTime {
+export interface TimeFromNowProps {
   time: Moment
 }
 
-export const TimeFromNow: React.FC<ItalicTime> = ({ time }) => {
+export const TimeFromNow: React.FC<TimeFromNowProps> = ({ time }) => {
   return (
     <time className={'mx-1'} title={time.format('LLLL')} dateTime={time.format()}>{time.fromNow(true)}</time>
   )
