+++ src/components/editor/document-bar/document-info/unitalic-bold-text.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export interface UnitalicBoldTextProps {
+  text: string ;
+}
+
+export const UnitalicBoldText: React.FC<UnitalicBoldTextProps> = ({ text }) => {
+  return <b className={'font-style-normal mr-1'}>{text}</b>
+}
