+++ src/style/index.scss
@@ -36,3 +36,7 @@ body {
 .overflow-y-scroll {
   overflow-y: scroll;
 }
+
+.font-style-normal {
+  font-style: normal;
+}
